Change amount of steps aliens move across the screen
Before they move down. This change is needed because we changed the
width of the screen. At the moment it's pretty gross because we have to
manually change this value (it's not automatically calculated).

At the moment the steps are placed in the alien manager to make it
easier to change.